Example Script to export CMP data to MS excel
Example to export data to excel, component references, partnames and packages are written in excel table.
This can script is a basic example how to export data to MS excel, you can update it to add more information for components or change it a little bit to export e.g. net information.